add saveDeploymentUrlToTmpConfigPath as a function
